Moved helper functions that require Mapbox API to their own file. Updated relevant imports.
